Refactor: Api instance 분리
